+++ cache-config/testing/ort-tests/tcdata/cachegroupsdeliveryservices.go
@@ -115,7 +115,9 @@ func (r *TCData) CreateTestCachegroupsDeliveryServices(t *testing.T) {
 		serverID := server.ID
 
 		serverDSes, _, err := TOSession.GetDeliveryServicesByServer(serverID)
-
+		if err != nil {
+			t.Fatalf("getting delivery services by server: %v", err)
+		}
 		for _, dsID := range dsIDs {
 			found := false
 			for _, serverDS := range serverDSes {